refactor: 기존 TextEditor.tsx 파일을 툴바 관리만 하는 ToolBar.tsx로 변경
